refactor(katana-core): refactor service module
commit-id:aac1f033
